requirements.txt
fastapi~=0.95.1
bcrypt~=4.0.1
python-dotenv~=1.0.0
motor~=3.1.2
pydantic~=1.10.7
pymongo~=4.3.3
python-slugify~=8.0.1